+++ inc/_ideas.php
@@ -1,8 +1,0 @@
-<?php 
-
-/**
- * This file is not in use. 
- * It is only here for ideas and probable future improvements.
- */
-
-if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly