+++ changelog.md
@@ -1,5 +1,6 @@
 #### 0.0.2 (2021-02-16)
 
+- deeper check for changed settings (#c6943ee)
 - check for 'title' support in most cases, canonical redirect for archives (#2b0020e)
 - add field setting bool `acfml_all_required` (#86fb2c3)
 - Remove ACF nag on `beforeunload` (#86fb2c3)